Move equipment model and related stuff to it's own module.
