+++ resources/models/__init__.py
@@ -1,5 +1,5 @@
 from .availability import Day, Period, get_opening_hours  # noqa
 from .reservation import Reservation  # noqa
-from .resource import Purpose, Resource, ResourceType, ResourceImage  # noqa
-from .resource import Equipment, ResourceEquipment, EquipmentAlias  # noqa
+from .resource import Purpose, Resource, ResourceType, ResourceImage, ResourceEquipment  # noqa
+from .equipment import Equipment, EquipmentAlias  # noqa
 from .unit import Unit, UnitIdentifier  # noqa
