+++ resources/models/resource.py
@@ -18,7 +18,8 @@ from django_hstore import hstore
 from resources.errors import InvalidImage
 
 from .base import AutoIdentifiedModel, ModifiableModel
-from .utils import get_translated, get_translated_name, DEFAULT_LANG
+from .utils import get_translated, get_translated_name
+from .equipment import Equipment
 
 
 class ResourceType(ModifiableModel, AutoIdentifiedModel):
@@ -58,18 +59,6 @@ class Purpose(ModifiableModel):
 
     def __str__(self):
         return "%s (%s)" % (get_translated(self, 'name'), self.id)
-
-
-class Equipment(ModifiableModel, AutoIdentifiedModel):
-    id = models.CharField(primary_key=True, max_length=100)
-    name = models.CharField(verbose_name=_('Name'), max_length=200)
-
-    class Meta:
-        verbose_name = _("equipment")
-        verbose_name_plural = _("equipment")
-
-    def __str__(self):
-        return get_translated(self, 'name')
 
 
 class Resource(ModifiableModel, AutoIdentifiedModel):
@@ -413,17 +402,3 @@ class ResourceEquipment(ModifiableModel):
 
     def __str__(self):
         return "%s / %s" % (self.equipment, self.resource)
-
-
-class EquipmentAlias(ModifiableModel, AutoIdentifiedModel):
-    id = models.CharField(primary_key=True, max_length=100)
-    name = models.CharField(verbose_name=_('Name'), max_length=200)
-    language = models.CharField(choices=settings.LANGUAGES, default=DEFAULT_LANG, max_length=3)
-    equipment = models.ForeignKey(Equipment, related_name='aliases')
-
-    class Meta:
-        verbose_name = _('equipment alias')
-        verbose_name_plural = _('equipment aliases')
-
-    def __str__(self):
-        return "%s (%s)" % (get_translated(self, 'name'), self.equipment)
